src/main/java/guru/springframework/recipe/app/converters/fromdomain/CategoryToCategoryCommand.java
package guru.springframework.recipe.app.converters.fromdomain;

import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

import guru.springframework.recipe.app.commands.CategoryCommand;
import guru.springframework.recipe.app.domain.Category;
import lombok.Synchronized;

@Component
public class CategoryToCategoryCommand implements Converter<Category, CategoryCommand> {

	@Synchronized
	@Nullable
	@Override
	public CategoryCommand convert(Category source) {
		if (source == null) {
			return null;
		}
		
		CategoryCommand destination = new CategoryCommand();
		destination.setId(source.getId());
		destination.setDescription(source.getDescription());
		return destination;
	}

}
